Handle symbols in image names
S3 events encode filenames both with URL encoding and by replacing spaces with plus signs. For example, `A file+name` is passed to Lambda as `A+file%2Bname`. However, S3 expects the original filename as the key. This causes aws-lambda-image to fail on images that have spaces or symbols in them.